+++ model_navigator/cli/helm_chart_create.py
@@ -16,6 +16,7 @@ import dataclasses
 import logging
 import traceback
 from pathlib import Path
+from typing import Optional
 
 import click
 
@@ -31,7 +32,9 @@ from model_navigator.cli.spec import (
     TritonModelSchedulerConfigCli,
 )
 from model_navigator.converter.config import ComparatorConfig, DatasetProfileConfig
+from model_navigator.framework import SUFFIX2FRAMEWORK
 from model_navigator.kubernetes import ChartGenerator, HelmChartGenerationResult
+from model_navigator.kubernetes.triton import TritonServer
 from model_navigator.log import init_logger, log_dict
 from model_navigator.model import ModelConfig, ModelSignatureConfig
 from model_navigator.results import ResultsStore, State, Status
@@ -42,6 +45,7 @@ from model_navigator.triton.config import (
 )
 from model_navigator.utils.cli import common_options, options_from_config
 from model_navigator.utils.workspace import Workspace
+from model_navigator.validators import run_command_validators
 
 LOGGER = logging.getLogger("helm_chart_create")
 
@@ -56,6 +60,7 @@ LOGGER = logging.getLogger("helm_chart_create")
     help="Path to Helm Charts repository.",
 )
 @click.option("--chart-name", required=True, help="Name of the chart in Helm Charts repository.")
+@click.option("--chart-version", required=False, help="Version of the chart in Helm Charts repository.")
 @options_from_config(ModelSignatureConfig, ModelSignatureConfigCli)
 @options_from_config(ConversionSetConfig, ConversionSetHelmChartConfigCli)
 @options_from_config(ComparatorConfig, ComparatorConfigCli)
@@ -71,11 +76,30 @@ def helm_chart_create_cmd(
     override_workspace: bool,
     charts_repository: str,
     chart_name: str,
+    chart_version: Optional[str],
     container_version: str,
+    triton_docker_image: Optional[str],
+    framework_docker_image: Optional[str],
     **kwargs,
 ):
     init_logger(verbose=verbose)
-    LOGGER.debug("Running helm_chart_create_cmd")
+    LOGGER.debug(f"Running '{ctx.command_path}' with config_path: {kwargs.get('config_path')}")
+
+    run_command_validators(
+        ctx.command.name,
+        configuration={
+            "verbose": verbose,
+            "workspace_path": workspace_path,
+            "override_workspace": override_workspace,
+            "charts_repository": charts_repository,
+            "chart_name": chart_name,
+            "chart_version": chart_version,
+            "container_version": container_version,
+            "triton_docker_image": triton_docker_image,
+            "framework_docker_image": framework_docker_image,
+            **kwargs,
+        },
+    )
 
     workspace = Workspace(workspace_path)
     charts_repository = Path(charts_repository).resolve()
@@ -88,6 +112,10 @@ def helm_chart_create_cmd(
     optimization_config = TritonModelOptimizationConfig.from_dict(kwargs)
     scheduler_config = TritonModelSchedulerConfig.from_dict(kwargs)
     instances_config = TritonModelInstancesConfig.from_dict(kwargs)
+
+    framework = SUFFIX2FRAMEWORK[src_model_config.model_path.suffix]
+    framework_docker_image = framework_docker_image or framework.container_image(container_version)
+    triton_docker_image = triton_docker_image or TritonServer.container_image(container_version)
 
     if verbose:
         log_dict(
@@ -106,6 +134,8 @@ def helm_chart_create_cmd(
                 "workspace_path": workspace.path,
                 "override_workspace": override_workspace,
                 "container_version": container_version,
+                "framework_docker_image": framework_docker_image,
+                "triton_docker_image": triton_docker_image,
                 "verbose": verbose,
             },
         )
@@ -117,8 +147,9 @@ def helm_chart_create_cmd(
     conversion_config = conversion_set_config_lst[0]
 
     try:
-        container_version_without_tag = container_version.split("-")[0]
-        helm_chart_generator = ChartGenerator(container_version=container_version_without_tag)
+        helm_chart_generator = ChartGenerator(
+            triton_docker_image=triton_docker_image, framework_docker_image=framework_docker_image
+        )
         helm_chart_generation_result = helm_chart_generator.run(
             src_model=src_model_config,
             src_model_signature_config=src_model_signature_config,
@@ -129,13 +160,15 @@ def helm_chart_create_cmd(
             scheduler_config=scheduler_config,
             instances_config=instances_config,
             output_path=output_path,
+            chart_version=chart_version,
         )
     except Exception:
         message = traceback.format_exc()
         LOGGER.debug(f"Encountered exception \n{message}")
         return HelmChartGenerationResult(
             status=Status(state=State.FAILED, message=message),
-            container_version=container_version,
+            triton_docker_image=triton_docker_image,
+            framework_docker_image=framework_docker_image,
             src_model_config=src_model_config,
             src_model_signature_config=src_model_signature_config,
             conversion_config=conversion_config,
@@ -148,6 +181,6 @@ def helm_chart_create_cmd(
         )
 
     results_store = ResultsStore(workspace)
-    results_store.dump("helm_chart_generation", [helm_chart_generation_result])
+    results_store.dump(ctx.command.name.replace("-", "_"), [helm_chart_generation_result])
 
     return helm_chart_generation_result