+++ model_navigator/kubernetes/inference/values.py
@@ -12,37 +12,36 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from model_navigator.kubernetes import helm
-from model_navigator.kubernetes.triton import TritonServer
 
 
 class Values(helm.Values):
-    def __init__(self, container_version: str):
-        self.container_version = container_version
+    def __init__(self, docker_image: str):
+        self.docker_image = docker_image
 
     def data(self):
-        values = dict(
-            imagePullSecret=None,
-            pullPolicy="Always",
-            restartPolicy="Always",
-            replicaCount=1,
-            gpu=dict(
-                limit=1,
-                product=None,
-                mig=None,
-            ),
-            deployer=dict(
-                image=None,
-                modelUri=None,
-                gcsCredentialsFile=None,
-                awsCredentialsFile=None,
-                azureCredentialsFile=None,
-            ),
-        )
-
-        values["server"] = {"image": TritonServer.container_image(version=self.container_version)}
-
-        values["service"] = {
-            "type": "ClusterIP",
+        values = {
+            "imagePullSecret": None,
+            "pullPolicy": "Always",
+            "restartPolicy": "Always",
+            "replicaCount": 1,
+            "gpu": {
+                "limit": 1,
+                "product": None,
+                "mig": None,
+            },
+            "deployer": {
+                "image": None,
+                "modelUri": None,
+                "gcsCredentialsFile": None,
+                "awsCredentialsFile": None,
+                "azureCredentialsFile": None,
+            },
+            "server": {
+                "image": self.docker_image,
+            },
+            "service": {
+                "type": "ClusterIP",
+            },
         }
 
         return values